Use delta time instead of frames
The math is now scaled so the game works mostly the same regardless of
what framerate it runs on.

Whenever we have a timer- for instance, the cayotee timer says that
you are still allowed to jump for five frames after having left the
ground- the timing is still done in frames, because we only check on
whole frames anyways, but the number of frames is scaled, then generously
rounded upwards. This means that you get more generous timers across
the board if your game runs a little slower